restructure atoms submodule in content package
