--- packages/content/component-layout.d.ts
@@ -1,1 +1,0 @@
-﻿export * from './src/component-layout/index';
