+++ packages/content/src/atoms/component-layout/document.ts
