+++ packages/content/src/atoms/component-layout/resolver.ts
