+++ packages/content/src/editing/index.ts
@@ -42,42 +42,6 @@ export {
   updateComponent,
 } from './design-library';
 export {
-  type Document as ComponentLayoutDocument,
-  type Node as ComponentLayoutNode,
-  type Element as ComponentLayoutElement,
-  type Binding as ComponentLayoutBinding,
-  type Primitive as ComponentLayoutPrimitive,
-  type ForLoop,
-  type ShowNode,
-  type ExpressionBinding as ComponentLayoutExpressionBinding,
-  type EventBinding as ComponentLayoutEventBinding,
-  type Action as ComponentLayoutAction,
-  type ShowComparison,
-  type ShowAnd,
-  type ShowOr,
-  isElement as isComponentLayoutElement,
-  hasFor as hasComponentLayoutFor,
-  hasShow as hasComponentLayoutShow,
-  isExpressionBinding as isComponentLayoutExpressionBinding,
-  isEventBinding as isComponentLayoutEventBinding,
-  isSetStateAction as isComponentLayoutSetStateAction,
-  isCallAction as isComponentLayoutCallAction,
-  isPrimitive as isComponentLayoutPrimitive,
-  isShowComparison as isComponentLayoutShowComparison,
-  isShowAnd as isComponentLayoutShowAnd,
-  isShowOr as isComponentLayoutShowOr,
-  type ResolveContext as ComponentLayoutResolveContext,
-  type ParsedBind as ComponentLayoutParsedBind,
-  type BindSource as ComponentLayoutBindSource,
-  type BindSegment as ComponentLayoutBindSegment,
-  parseBindExpression as parseComponentLayoutBindExpression,
-  resolveBindExpression as resolveComponentLayoutBindExpression,
-  isTemplateString as isComponentLayoutTemplateString,
-  resolveTemplateString as resolveComponentLayoutTemplateString,
-  evaluateShowNode as evaluateComponentLayoutShowNode,
-  resolveIfTemplate,
-} from '../component-layout';
-export {
   AtomInfo,
   CallbackInfo,
   AtomType,
